requirements.txt
# Core dependencies
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.3
numpy>=1.20.0
matplotlib>=3.4.0
pillow>=8.2.0

# Data processing
albumentations>=1.0.0
scikit-learn>=0.24.0
pandas>=1.3.0

# Visualization
seaborn>=0.11.0

# Utils
tqdm>=4.62.0
pyyaml>=5.4.0

# Optional: GPU support
# Uncomment if using CUDA
# cudatoolkit>=11.1
